+++ tests/unit/IExecOracleFactory/createOracle.test.ts
@@ -516,7 +516,7 @@ describe('createOracle', () => {
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
     expect(errors[0].message).toBe('Failed to upload paramSet');
-    expect(errors[0].originalError).toStrictEqual(Error('ipfs.add failed'));
+    expect(errors[0].cause).toStrictEqual(Error('ipfs.add failed'));
   }, 10000);
 
   test('error - unexpected error', async () => {
@@ -553,8 +553,8 @@ describe('createOracle', () => {
     expect(errors.length).toBe(1);
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
-    expect(errors[0].message).toBe('Create oracle unexpected error');
-    expect(errors[0].originalError).toBeInstanceOf(TypeError);
+    expect(errors[0].message).toBe('Failed to create oracle');
+    expect(errors[0].cause).toBeInstanceOf(TypeError);
   }, 10000);
 
   test('error - with apiKey failed to encrypt apiKey', async () => {
@@ -599,7 +599,7 @@ describe('createOracle', () => {
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
     expect(errors[0].message).toBe('Failed to encrypt API key');
-    expect(errors[0].originalError).toStrictEqual(
+    expect(errors[0].cause).toStrictEqual(
       Error('iexec.dataset.encrypt failed')
     );
   }, 10000);
@@ -651,7 +651,7 @@ describe('createOracle', () => {
     expect(errors[0].message).toBe(
       'Failed to compute encrypted API key checksum'
     );
-    expect(errors[0].originalError).toStrictEqual(
+    expect(errors[0].cause).toStrictEqual(
       Error('iexec.dataset.computeEncryptedFileChecksum failed')
     );
   }, 10000);
@@ -692,7 +692,7 @@ describe('createOracle', () => {
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
     expect(errors[0].message).toBe('Failed to upload encrypted API key');
-    expect(errors[0].originalError).toStrictEqual(Error('ipfs.add failed'));
+    expect(errors[0].cause).toStrictEqual(Error('ipfs.add failed'));
   }, 10000);
 
   test('error - with apiKey failed to deploy dataset', async () => {
@@ -740,8 +740,10 @@ describe('createOracle', () => {
     expect(messages.length).toBe(4);
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
-    expect(errors[0].message).toBe('Failed to deploy API key dataset');
-    expect(errors[0].originalError).toStrictEqual(
+    expect(errors[0].message).toBe(
+      'Failed to create dataset containing encrypted API key'
+    );
+    expect(errors[0].cause).toStrictEqual(
       Error('iexec.dataset.deployDataset failed')
     );
   }, 10000);
@@ -803,9 +805,9 @@ describe('createOracle', () => {
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
     expect(errors[0].message).toBe(
-      "Failed to push API key dataset's encryption key"
+      'Failed to create dataset containing encrypted API key'
     );
-    expect(errors[0].originalError).toStrictEqual(
+    expect(errors[0].cause).toStrictEqual(
       Error('iexec.dataset.pushDatasetSecret failed')
     );
   }, 10000);
@@ -872,8 +874,8 @@ describe('createOracle', () => {
     expect(messages.length).toBe(7);
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
-    expect(errors[0].message).toBe("Failed to create API key datasetorder's");
-    expect(errors[0].originalError).toStrictEqual(
+    expect(errors[0].message).toBe('Failed to create API key dataset order');
+    expect(errors[0].cause).toStrictEqual(
       Error('iexec.order.createDatasetorder failed')
     );
   }, 10000);
@@ -955,7 +957,7 @@ describe('createOracle', () => {
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
     expect(errors[0].message).toBe("Failed to sign API key datasetorder's");
-    expect(errors[0].originalError).toStrictEqual(
+    expect(errors[0].cause).toStrictEqual(
       Error('iexec.order.signDatasetorder failed')
     );
   }, 10000);
@@ -1000,9 +1002,9 @@ describe('createOracle', () => {
     expect(messages.length).toBe(0);
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
-    expect(errors[0].message).toBe('API key dataset creation unexpected error');
-    expect(errors[0].originalError).toStrictEqual(
-      Error('something bad happened')
+    expect(errors[0].message).toBe(
+      'Failed to create dataset containing encrypted API key'
     );
+    expect(errors[0].cause).toStrictEqual(Error('something bad happened'));
   }, 10000);
 });