+++ tests/unit/IExecOracleFactory/updateOracle.test.ts
@@ -627,7 +627,7 @@ describe('updateOracle', () => {
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
     expect(errors[0].message).toBe('Failed to load paramSet');
-    expect(errors[0].originalError).toStrictEqual(
+    expect(errors[0].cause).toStrictEqual(
       Error(
         'Failed to load paramSetSet from CID QmTJ41EuPEwiPTGrYVPbXgMGvmgzsRYWWMmw6krVDN94nh'
       )
@@ -666,7 +666,7 @@ describe('updateOracle', () => {
     expect(errors.length).toBe(1);
     // expect(errors[0]).toBeInstanceOf(WorkflowError);
     expect(errors[0].message).toBe('Failed to load paramSet');
-    // expect(errors[0].originalError).toStrictEqual(
+    // expect(errors[0].cause).toStrictEqual(
     //   Error(
     //     'Content associated to CID QmTJ41EuPEwiPTGrYVPbXgMGvmgzsRYWWMmw6krVDN94nh is not a valid paramSet',
     //   ),
@@ -742,7 +742,7 @@ describe('updateOracle', () => {
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
     expect(errors[0].message).toBe('Failed to upload paramSet');
-    expect(errors[0].originalError).toStrictEqual(Error('ipfs.add failed'));
+    expect(errors[0].cause).toStrictEqual(Error('ipfs.add failed'));
   }, 10000);
 
   test('error - fail to fetch apporder', async () => {
@@ -789,8 +789,8 @@ describe('updateOracle', () => {
     expect(messages.length).toBe(4);
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
-    expect(errors[0].message).toBe('Failed to fetch apporder');
-    expect(errors[0].originalError).toStrictEqual(
+    expect(errors[0].message).toBe('Failed to update oracle');
+    expect(errors[0].cause).toStrictEqual(
       Error('iexec.orderbook.fetchAppOrderbook failed')
     );
   }, 10000);
@@ -837,8 +837,8 @@ describe('updateOracle', () => {
     expect(messages.length).toBe(4);
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
-    expect(errors[0].message).toBe('No apporder published');
-    expect(errors[0].originalError).toBeUndefined();
+    expect(errors[0].message).toBe('Failed to update oracle');
+    expect(errors[0].cause).toStrictEqual(Error('No app order published'));
   }, 10000);
 
   test('error - fail to fetch datasetorder', async () => {
@@ -889,8 +889,8 @@ describe('updateOracle', () => {
     expect(messages.length).toBe(6);
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
-    expect(errors[0].message).toBe('Failed to fetch datasetorder');
-    expect(errors[0].originalError).toStrictEqual(
+    expect(errors[0].message).toBe('Failed to update oracle');
+    expect(errors[0].cause).toStrictEqual(
       Error('iexec.orderbook.fetchDatasetOrderbook fail')
     );
   }, 10000);
@@ -941,11 +941,11 @@ describe('updateOracle', () => {
     expect(messages.length).toBe(6);
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
-    expect(errors[0].message).toBe('No datasetorder published');
-    expect(errors[0].originalError).toBeUndefined();
+    expect(errors[0].message).toBe('Failed to update oracle');
+    expect(errors[0].cause).toStrictEqual(Error('No dataset order published'));
   }, 10000);
 
-  test('error - fail to fetch workerppolorder', async () => {
+  test('error - fail to fetch workerpool order', async () => {
     mockAdd.mockResolvedValueOnce(
       'QmTJ41EuPEwiPTGrYVPbXgMGvmgzsRYWWMmw6krVDN94nh'
     ) as any;
@@ -997,8 +997,8 @@ describe('updateOracle', () => {
     expect(messages.length).toBe(8);
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
-    expect(errors[0].message).toBe('Failed to fetch workerpoolorder');
-    expect(errors[0].originalError).toStrictEqual(
+    expect(errors[0].message).toBe('Failed to update oracle');
+    expect(errors[0].cause).toStrictEqual(
       Error('iexec.orderbook.fetchWorkerpoolOrderbook fail')
     );
   }, 10000);
@@ -1053,8 +1053,10 @@ describe('updateOracle', () => {
     expect(messages.length).toBe(8);
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
-    expect(errors[0].message).toBe('No workerpoolorder published');
-    expect(errors[0].originalError).toBeUndefined();
+    expect(errors[0].message).toBe('Failed to update oracle');
+    expect(errors[0].cause).toStrictEqual(
+      Error('No workerpool order published')
+    );
   }, 10000);
 
   test('error - fail to create requestorder', async () => {
@@ -1113,8 +1115,8 @@ describe('updateOracle', () => {
     expect(messages.length).toBe(9);
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
-    expect(errors[0].message).toBe('Failed to create requestorder');
-    expect(errors[0].originalError).toStrictEqual(
+    expect(errors[0].message).toBe('Failed to create request order');
+    expect(errors[0].cause).toStrictEqual(
       Error('iexec.order.createRequestorder failed')
     );
   }, 10000);
@@ -1179,7 +1181,7 @@ describe('updateOracle', () => {
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
     expect(errors[0].message).toBe('Failed to sign requestorder');
-    expect(errors[0].originalError).toStrictEqual(
+    expect(errors[0].cause).toStrictEqual(
       Error('iexec.order.signRequestorder failed')
     );
   }, 10000);
@@ -1246,8 +1248,8 @@ describe('updateOracle', () => {
     expect(messages.length).toBe(12);
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
-    expect(errors[0].message).toBe('Failed to match orders');
-    expect(errors[0].originalError).toStrictEqual(
+    expect(errors[0].message).toBe('Failed to update oracle');
+    expect(errors[0].cause).toStrictEqual(
       Error('iexec.order.matchOrders failed')
     );
   }, 10000);
@@ -1322,9 +1324,7 @@ describe('updateOracle', () => {
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
     expect(errors[0].message).toBe('Failed to monitor oracle update task');
-    expect(errors[0].originalError).toStrictEqual(
-      Error('iexec.task.obsTask error')
-    );
+    expect(errors[0].cause).toStrictEqual(Error('iexec.task.obsTask error'));
   }, 10000);
 
   test('error - update task timedout', async () => {
@@ -1400,7 +1400,7 @@ describe('updateOracle', () => {
     expect(errors[0].message).toBe(
       'Oracle update task timed out, update failed'
     );
-    expect(errors[0].originalError).toStrictEqual(
+    expect(errors[0].cause).toStrictEqual(
       Error('Task taskid from deal dealid timed out')
     );
   }, 10000);
@@ -1444,7 +1444,7 @@ describe('updateOracle', () => {
     expect(messages.length).toBe(2);
     expect(errors.length).toBe(1);
     expect(errors[0]).toBeInstanceOf(WorkflowError);
-    expect(errors[0].message).toBe('Update oracle unexpected error');
-    expect(errors[0].originalError).toBeInstanceOf(TypeError);
+    expect(errors[0].message).toBe('Failed to update oracle');
+    expect(errors[0].cause).toBeInstanceOf(TypeError);
   }, 10000);
 });