+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
+## Next
+
+### Changed
+
+- Upgraded to iexec ^8.9.1
+- [BREAKING] Ship ES2022 JavaScript instead of es2015 (aka es6) in order to support `errorCause` optional field in `Error`:
+  - Minimum browser versions: <https://gist.github.com/Julien-Marcou/156b19aea4704e1d2f48adafc6e2acbf>
+  - Minimum Node.js version: 18
+- Upgrade typescript version
+- Changed `fetchMyContacts`, `fetchUserContacts` and `sendEmail` error handling:
+  - Distinguish iExec protocol errors from other errors
+  - Store original error as the error errorCause
+- [BREAKING] Removed `originalError` from `WorkflowError`
+
 ## [2.1.0] 2024-03-15
 
 ### Removed